+++ src/Console/TubePeekCommand.php
@@ -23,14 +23,29 @@ class TubePeekCommand extends AbstractCommand
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $status = $input->getOption('status');
+
+        // Fail early if status is not supported
         if (!in_array($status, ['ready', 'delayed', 'buried'], true)) {
             throw new InvalidArgumentException("Specified status '{$status}' is not valid.");
         }
 
         $this->getBeanstalk()
             ->useTube($input->getArgument('tube'));
-        $method = 'peek' . ucfirst($status);
-        $job = call_user_func([$this->getBeanstalk(), $method]);
+
+        // Use switch instead of `->{'peek' . $status}` to allow static analysis
+        switch ($status) {
+            case 'ready':
+                $job = $this->getBeanstalk()->peekReady();
+                break;
+            case 'delayed':
+                $job = $this->getBeanstalk()->peekDelayed();
+                break;
+            case 'buried':
+                $job = $this->getBeanstalk()->peekBuried();
+                break;
+            default:
+                throw new InvalidArgumentException("Specified status '{$status}' is not valid.");
+        }
 
         if ($job === false) {
             $output->writeln("No jobs found in '{$status}' status.");